Deduplicate CanteraYamlFileComparer into a shared test helper.
yaml_cantera1Test.py and yaml_cantera2Test.py each carried a ~145-line
copy of CanteraYamlFileComparer, a pytest fixture-based base class that
asserts equivalence between two Cantera YAML files (generator, phases,
elements, species composition/thermo/transport). 

Move the class into a new sibling helper module
test/rmgpy/cantera_yaml_comparer.py and import it from both test files.
