+++ test/rmgpy/yaml_cantera1Test.py
@@ -31,12 +31,11 @@
 Tests for rmgpy.yaml_cantera1 module.
 """
 
-import copy
 import os
 import pytest
 import numpy as np
-import yaml
 
+from cantera_yaml_comparer import CanteraYamlFileComparer
 from rmgpy.species import Species
 from rmgpy.reaction import Reaction
 from rmgpy.thermo import NASA, NASAPolynomial
@@ -551,155 +550,6 @@ class TestCanteraWriter1:
         """Test that CanteraWriter1 can be instantiated."""
         writer = CanteraWriter1()
         assert writer is not None
-
-class CanteraYamlFileComparer:
-    """
-    For comparing two Cantera YAML files. 
-    This class provides methods to compare species and reactions between the two files.
-
-        Args:
-            yaml_path_1: Path to the first YAML file, converted from Chemkin by ck2yaml.
-            yaml_path_2: Path to the second YAML file, written directly by RMG.
-    """
-    yaml_path_1 = None
-    yaml_path_2 = None
-
-    @pytest.fixture(autouse=True, scope="class") # loaded once per Class
-    def load_yaml_files(self, request):
-        """Load the two YAML files to be compared."""
-        with open(request.cls.yaml_path_1, 'r') as file:
-            request.cls.yaml1 = yaml.safe_load(file)
-        with open(request.cls.yaml_path_2, 'r') as file:
-            request.cls.yaml2 = yaml.safe_load(file)
-
-    @pytest.fixture(autouse=True)  # runs before each test method
-    def copy_yaml_dicts(self):
-        """Make deep copies so tests can modify without affecting other tests."""
-        self.yaml1 = copy.deepcopy(self.__class__.yaml1)
-        self.yaml2 = copy.deepcopy(self.__class__.yaml2)
-
-    def testGeneratorsAsExpected(self):
-        "Check the two yaml files were generated by the expected tools (ck2yaml vs RMG)."
-        assert self.yaml1['generator'] == 'ck2yaml', "First YAML file should be generated by ck2yaml."
-        assert 'RMG' in self.yaml2['generator'], "Second YAML file should be generated by RMG."
-
-    def testKeysMatch(self):
-        """Test that the top-level keys in both YAML files match, except those expected not to."""
-        # Remove keys from ck2yaml output that are not present in RMG output
-        self.yaml1.pop('input-files', None)
-        self.yaml1.pop('cantera-version', None)
-        for model in [self.yaml1, self.yaml2]:
-            for phase in model['phases']:
-                for reactions_block in phase.get('reactions', []): # for multi-phase mechanisms, reactions are under each phase
-                    assert reactions_block in model, f"Expected reactions block '{reactions_block}' not found in YAML file."
-                    model.pop(reactions_block, None)  # Remove reactions block to allow keys to match
-        assert self.yaml1.keys() == self.yaml2.keys(), "YAML files have different top-level keys."
-    
-    def testPhasesMatch(self):
-        """Test that the phase definitions in both YAML files match."""
-        assert len(self.yaml1['phases']) == len(self.yaml2['phases']), "YAML files have different numbers of phases"
-
-        for phase1, phase2 in zip(self.yaml1['phases'], self.yaml2['phases']):
-            assert phase1['name'] == phase2['name'], f"Phase names do not match: {phase1['name']} vs {phase2['name']}."
-            assert phase1['thermo'] == phase2['thermo'], f"Thermo definitions for phase {phase1['name']} do not match."
-            assert phase1.get('transport', '') == phase2.get('transport', ''), f"Transport definitions for phase {phase1['name']} do not match."
-            assert phase1.get('adjacent-phases', []) == phase2.get('adjacent-phases', []), f"Adjacent phases for phase {phase1['name']} do not match."
-            assert phase1.get('species', []) == phase2.get('species', []), f"Species lists for phase {phase1['name']} do not match."
-            assert phase1.get('reactions', []) == phase2.get('reactions', []), f"Reactions blocks for phase {phase1['name']} do not match."
-            # the ck2yaml has all elements in Titlecase, while RMG lets some isotopes be CI and OI (not Ci and Oi).
-            assert sorted(phase1.get('elements', [])) == sorted(e.title() for e in phase2.get('elements', [])), f"Element lists for phase {phase1['name']} do not match."
-            assert phase1.get('state', {}) == phase2.get('state', {}), f"State definitions for phase {phase1['name']} do not match."
-
-    def testElementsMatch(self):
-        """Test that the element definitions in both YAML files match."""
-        assert ('elements' in self.yaml1) == ('elements' in self.yaml2), "One YAML file has 'elements' block while the other does not."
-        ck2yaml_elements = sorted(self.yaml1.get('elements', []), key=lambda e: e['symbol'])
-        # Put symbol into Titlecase to match ck2yaml's formatting
-        rmg_elements = [{'symbol': e['symbol'].title(), 'atomic-weight': e['atomic-weight']} for e in self.yaml2.get('elements', [])]
-        # Sort by the 'symbol' key.
-        rmg_elements = sorted(rmg_elements, key=lambda e: e['symbol'])
-        # Compare symbols exactly, and atomic weights approximately
-        assert [e['symbol'] for e in ck2yaml_elements] == [e['symbol'] for e in rmg_elements], \
-            "YAML files have different element symbols."
-        assert [e['atomic-weight'] for e in ck2yaml_elements] == pytest.approx(
-            [e['atomic-weight'] for e in rmg_elements], abs=1e-3
-        ), "YAML files have different element atomic weights."
-
-    def testSpeciesMatch(self):
-        """Test that species definitions match between the two YAML files."""
-        species1 = {s['name']: s for s in self.yaml1['species']}
-        species2 = {s['name']: s for s in self.yaml2['species']}
-        assert species1.keys() == species2.keys(), "Species names do not match."
-
-        for name in species1:
-            s1 = species1[name]
-            s2 = species2[name]
-
-            # Composition: ck2yaml uses int values, RMG uses float
-            assert {k: int(v) for k, v in s2['composition'].items()} == s1['composition'], \
-                f"Composition mismatch for {name}."
-
-            # Thermo model
-            assert s1['thermo']['model'] == s2['thermo']['model'], \
-                f"Thermo model mismatch for {name}."
-
-            # Temperature ranges and polynomial data
-            # ck2yaml may collapse single-polynomial NASA7 (e.g. Ar) into one range
-            # while RMG always writes two polynomials with a midpoint temperature.
-            t_ranges1 = s1['thermo'].get('temperature-ranges', [])
-            t_ranges2 = s2['thermo'].get('temperature-ranges', [])
-            data1 = s1['thermo'].get('data', [])
-            data2 = s2['thermo'].get('data', [])
-
-            if len(t_ranges1) == 2 and len(t_ranges2) == 3:
-                # ck2yaml collapsed to single polynomial; RMG has two identical ones
-                assert t_ranges1[0] == pytest.approx(t_ranges2[0], rel=1e-4), \
-                    f"Temperature range lower bound mismatch for {name}."
-                assert t_ranges1[1] == pytest.approx(t_ranges2[2], rel=1e-4), \
-                    f"Temperature range upper bound mismatch for {name}."
-                assert len(data1) == 1 and len(data2) == 2, \
-                    f"Expected 1 vs 2 polynomials for collapsed species {name}."
-                assert data1[0] == pytest.approx(data2[0], rel=1e-4), \
-                    f"Thermo polynomial mismatch for {name} (low range)."
-                assert data1[0] == pytest.approx(data2[1], rel=1e-4), \
-                    f"Thermo polynomial mismatch for {name} (high range should match low)."
-            else:
-                assert t_ranges1 == pytest.approx(t_ranges2, rel=1e-4), \
-                    f"Temperature ranges mismatch for {name}."
-                assert len(data1) == len(data2), \
-                    f"Number of thermo polynomial ranges differs for {name}."
-                for i, (poly1, poly2) in enumerate(zip(data1, data2)):
-                    assert poly1 == pytest.approx(poly2, rel=1e-4), \
-                        f"Thermo polynomial {i} mismatch for {name}."
-            # Ideally thermo data would have notes.
-
-            # RMG includes reference-pressure but ck2yaml does not (when it's non-default)
-            # (no assertion needed, just noting the known difference)
-
-            # Transport data
-            assert ('transport' in s1) == ('transport' in s2), f"Transport data presence mismatch for {name}."
-            if 'transport' in s1 and 'transport' in s2:
-                t1 = s1['transport']
-                t2 = s2['transport']
-                assert t1['model'] == t2['model'], f"Transport model mismatch for {name}."
-                assert t1['geometry'] == t2['geometry'], f"Transport geometry mismatch for {name}."
-                assert t1.get('well-depth', 0) == pytest.approx(
-                    t2.get('well-depth', 0), rel=1e-3
-                ), f"Transport well-depth mismatch for {name}."
-                assert t1.get('diameter', 0) == pytest.approx(
-                    t2.get('diameter', 0), rel=1e-3
-                ), f"Transport diameter mismatch for {name}."
-                assert t1.get('polarizability', 0) == pytest.approx(
-                    t2.get('polarizability', 0), rel=1e-3
-                ), f"Transport polarizability mismatch for {name}."
-                assert t1.get('dipole', 0) == pytest.approx(
-                    t2.get('dipole', 0), rel=1e-3
-                ), f"Transport dipole mismatch for {name}."
-                assert t1.get('rotational-relaxation', 0) == pytest.approx(
-                    t2.get('rotational-relaxation', 0), rel=1e-3
-                ), f"Transport rotational-relaxation mismatch for {name}."
-                assert t1.get('note', '') == t2.get('note', ''), \
-                    f"Transport note mismatch for {name}."
 
 @pytest.mark.skip(reason="These files are out of date and have been removed.")
 class TestPreviouslyWrittenCanteraYamlGasOnly(CanteraYamlFileComparer):
